refactor: implement EXCSTurnoutsManager for managing turnouts in EX-CommandStation
